Remove the simplifyMergeNodes step
simplifyMergeNodes() was called to remove apparently useless decision
nodes but it appears that it corrupt the "predecessors" attribute fields
of PHI nodes, in spite of many attempts to fix it. We remove that step
for now in the releases. It may come back later when the problem is
identified and fixed.